Desktop/webapp-final/requirements.txt
python-telegram-bot==20.7
gspread==5.12.4
google-auth==2.27.0
python-dotenv==1.0.0
pytz==2024.1
flask==3.0.3
flask-cors==4.0.1
requests==2.31.0
